+++ redacted-compiler-plugin/src/main/kotlin/dev/zacsweers/redacted/compiler/fir/RedactedFirBuiltIns.kt
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.redacted.compiler.fir
 
+import dev.zacsweers.redacted.compiler.RedactionPlanRegistry
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.extensions.FirExtensionSessionComponent
 import org.jetbrains.kotlin.name.ClassId
@@ -10,12 +11,24 @@ internal class RedactedFirBuiltIns(
   session: FirSession,
   val redactedAnnotations: Set<ClassId>,
   val unRedactedAnnotations: Set<ClassId>,
+  val replacementString: String,
+  val planRegistry: RedactionPlanRegistry,
 ) : FirExtensionSessionComponent(session) {
   companion object {
-    fun getFactory(redactedAnnotations: Set<ClassId>, unRedactedAnnotations: Set<ClassId>) =
-      Factory { session ->
-        RedactedFirBuiltIns(session, redactedAnnotations, unRedactedAnnotations)
-      }
+    fun getFactory(
+      redactedAnnotations: Set<ClassId>,
+      unRedactedAnnotations: Set<ClassId>,
+      replacementString: String,
+      planRegistry: RedactionPlanRegistry,
+    ) = Factory { session ->
+      RedactedFirBuiltIns(
+        session,
+        redactedAnnotations,
+        unRedactedAnnotations,
+        replacementString,
+        planRegistry,
+      )
+    }
   }
 }
 
@@ -27,3 +40,9 @@ internal val FirSession.redactedAnnotations: Set<ClassId>
 
 internal val FirSession.unRedactedAnnotations: Set<ClassId>
   get() = redactedFirBuiltIns.unRedactedAnnotations
+
+internal val FirSession.redactedReplacementString: String
+  get() = redactedFirBuiltIns.replacementString
+
+internal val FirSession.redactionPlanRegistry: RedactionPlanRegistry
+  get() = redactedFirBuiltIns.planRegistry